[jetpack env] Consolidate EnvStore.List and EnvStore.Get, and minor cleanups (#2158)
* [jetpack env] list command should work for more than 10 env-vars

* [jetpack env] Consolidate EnvStore.List and EnvStore.Get, and minor cleanup

1. Consolidate EnvStore.List and EnvStore.Get
Previously, EnvStore.List was used to show the "secrets" during `jetpack secrets list`.
And `EnvStore.Get` was used during `jetpack up`, where a user was required to specify
the specific names of the secrets they wanted to use from their org for deploying their app.

Now that we are changing the scope to be project specific, during `jetpack up`,
we'll include all the secrets that have the projectID tag. So, we no longer need
this distinction between List and Get.

This PR consolidates it into Get.

2. Minor cleanup
- made some variable names shorter, like `configuration` -> `config`.
- rename `EnvStore.DeleteEnvironmentSecrets` -> `EnvStore.Delete`
- tried to remove "secret" from error-messages, print statements and variable names.
These are now less appropriate given the focus on env-vars for the product.