[envsec] Remove kind: ENVIRONMENT_VARIABLE (#2170)
## Summary
We used to tag secrets with a `kind` tag to indicate whether the secret was a mounted file or an environment variable. Since we no longer support mounting files, this PR removes the `kind` tag altogether to simplify the code.

## How was it tested?
Re-compiled.

## Is this change backwards-compatible?
Yes, because no one was using the secrets mount functionality.