[jetconfig] Allow url in jetconfig to be set by environment (#2694)
## Summary

This allows url to be set by environment. So

```yaml
url: www.example.com
```
and

```yaml
url: 
  staging: staging.example.com
  prod: prod.example.com
```

are both valid.

If url is string, it is only used for prod. Otherwise we use the current
environment to select the right value (possibly nil)

Not sure if this is best solution, but it helps unblock Castor. It is
also easily extensible to other fields with minimal work.

## How was it tested?

Deployed quick.yaml with following values:

```yaml
url: none.test.jtpck.com
```

```yaml
url: 
  dev: dev.test.jtpck.com
  staging: staging.test.jtpck.com
```

Deployed each case in dev, prod, and staging Tested all endpoints.

## Is this change backwards-compatible?
YES
Former-commit-id: 8319b6916722c9b4af9dd02cc9041da7d637b41d