[envsec] Implement logged-in experience, change varPath (#116)
## Summary

This handles generating SSMConfig and EnvID when user is logged in. It
continues to work as expected when user is not logged in if user has
provided org-id and project-id flags.

It changes varPaths to be:

`<org-id>/<project-id>/<env>/<name>` We can allow customization of this
path in the future.

It continues to use tags to find the env vars. One optimization we can
do is what @mohsenari suggested, if the user does not change the custom
varPath, we can use `GetByPath` which saves us an API call.

## How was it tested?

tested `envsec ls` without logging in, with logging in, without project.
Verified correct error message.

Was not able to test end-to-end because federated permissions are not
fully set up. So I get (expected) error:

```
Error: operation error SSM: DescribeParameters, https response error StatusCode: 400, RequestID: 35e504b6-8b86-4b47-b169-4384f50eb31b, api error AccessDeniedException: User: arn:aws:sts::984256416385:assumed-role/JetpackAssumeFederatedIdentityRole/CognitoIdentityCredentials is not authorized to perform: ssm:DescribeParameters on resource: arn:aws:ssm:us-west-2:984256416385:* because no identity-based policy allows the ssm:DescribeParameters action
```