Use the new auth library from envsec (#133)
## Summary
Replaces the authentication in the envsec cli by the new auth library,
allowing us to use Stytch tokens. Note that as written the code is
pulling the auth library from `sandbox` which is *not* being published
separately (so technically we're breaking the `envsec` repo if you try
to build from there).

I thought this might be ok since we're not yet publishing envsec. And
the next PR will publish the auth library and fix this. However, let me
know if you'd like me to change the order of operations.

## How was it tested?
Ran locally