[envsec] Add List function to envsec library (#252)
## Summary

This adds a `List` and `PrintEnvVars` functions to the envsec library.
It allows a 3rd party (e.g. devbox) app to use envsec functionality more
easily.

I refactored stores as much as I could to make this easy to use, but
it's still a bit tangled. Specifically, the envsec package uses the
jetpack API internally a bunch. In theory, we could make most of these
functions part of the `store`. That way other stores can implement their
own `init` or `whoami` specifics depending on their capabilities. This
is left for later.

Once this is merged, devbox will be able to:

* Call `envsec.List` in order to fetch secrets (instead of the current
implementation which calls the envsec CLI).
* Implement a `devbox secrets list` command that fetches and prints
secrets.

## How was it tested?

```bash
envsec ls
envsec foo=123
envsec ls
```