Replace go.jetpack.io with go.jetify.com (#435)
## Community Contribution License
All community contributions in this pull request are licensed to the
project maintainers
under the terms of the [Apache 2
license](https://www.apache.org/licenses/LICENSE-2.0).
By creating this pull request I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 license.

## Summary
Migrate to go.jetify.com in the opensource repo. As expected, after I
merge this in, everything will break ... but I need to check it in
first, before I can update all the go.mods.

Plan is as follows:
- [ ] Merge this PR
- [ ] Run go tidy inside opensource
- [ ] Update imports in axiom, and run go tidy
- [ ] Bump major version of any opensource packages we are officially
releasing (like typeid). If we're not doing tagged releases, I'm gonna
assume people shouldn't be depending directly on HEAD, so I'm ok
breaking there.

## How was it tested?
Wasn't.