Run go get -u ./... in the different modules (#439)
## Community Contribution License
All community contributions in this pull request are licensed to the
project maintainers
under the terms of the [Apache 2
license](https://www.apache.org/licenses/LICENSE-2.0).
By creating this pull request I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 license.

## Summary
Run go get -u ./... in the different modules.
Because the issue we're having in axiom is around dependencies, I'm
trying to upgrade all dependencies. Starting on the opensource side.

## How was it tested?
devbox run test