[deps] Update devbox and some go deps (#450)
## Summary

## How was it tested?

## Community Contribution License

All community contributions in this pull request are licensed to the
project
maintainers under the terms of the
[Apache 2 License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 License as
stated in
the
[Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Chores**
- Updated numerous external dependencies to their latest versions,
enhancing stability, security, and overall performance.
- Added new indirect dependencies to support additional features and
improvements.
- Introduced a new shell script for managing Python virtual
environments, improving setup and validation processes.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->